+++ oras/utils/__init__.py
@@ -13,6 +13,7 @@ from .fileio import (
     read_json,
     readline,
     recursive_find,
+    workdir,
     write_file,
     write_json,
 )